added method for writing to file, fixed check_password method

+++ bank_account_project/bank_account.rb
@@ -1,3 +1,5 @@
+# Final Project for Rampup Ruby class @ Startup Institute 
+# Jen Trudell May 11, 2015
 
 require './bank_module'
 require 'yaml'
@@ -6,45 +8,48 @@ def prompt
 	print ">> "
 end
 
-# method to open a new account and write it to customer's yaml file
+# method to open a new account and write BankAccount object to customer's yaml file
 def open_account
 include Bank
-	puts "What is the customer's name?"
+	puts "Welcome to the First Bank of Ruby! We are glad you have decided to bank with us."
+	puts "Please enter your first name: "
 	prompt
 	customer_name = gets.chomp
-	puts "What is the new account number?"
-	prompt
-	account_number = gets.chomp.to_i
-	puts "What is the opening balance of the account?"
+	puts "Enter the opening balance for your new account: "
 	prompt
 	opening_balance = gets.chomp.to_f
-	puts "Enter a password for the customer."
+	puts "Choose a password for your account: "
 	prompt
 	customer_password = gets.chomp
+	account_number = rand(1..1000)
+	puts "Congratulations! You have opened an account at First Bank of Ruby."
+	puts "Your account number is #{account_number}."
 	new_customer = BankAccount.new(customer_name, account_number, opening_balance, customer_password)
 	# write the new_customer object to the customer's yaml file
-	customer_file = "#{customer_name}" + ".yaml"
-	File.open(customer_file, 'w') do |file|  
-  		file.puts YAML::dump(new_customer)
-	end  
+	new_customer.update_record(customer_name)
+
 end
 
-
-
-# method to load account from customer's yaml file
+# method to load BankAccount object from customer's yaml file
 def load_account
 include Bank
-	puts "What customer's account would you like to load?"
+	puts "Please enter your first name."
 	prompt
 	customer_name = gets.chomp
 	customer_file = "#{customer_name}" + ".yaml"
-	File.open(customer_file, 'r') do |file|
-  	YAML::load(file.read)
+	# if file with the customer's name exists, loads the BankAccount object from the file
+	if File.exists?(customer_file)
+		File.open(customer_file, 'r') do |file|
+	  	YAML::load(file.read)
+	  end
+	# if customer doesn't exist, calls open_account to create a file/account for the customer, and then calls load_account
+	else
+		puts "It doesn't look like you have an account at First Bank of Ruby."
+		puts "Let's open a new account for you."
+		open_account
+		load_account
 	end 
 end
 
-
-#open_account
-test_account = load_account
-test_account.transactions
-
+# calls welcome method on BankAccount object that was loaded
+load_account.welcome
